+++ src/PhpSpreadsheet/Worksheet/Worksheet.php
@@ -35,6 +35,7 @@ use PhpOffice\PhpSpreadsheet\Style\Style;
 use PhpOffice\PhpSpreadsheet\Worksheet\Sparkline\Sparkline;
 use PhpOffice\PhpSpreadsheet\Worksheet\Sparkline\SparklineGroup;
 use PhpOffice\PhpSpreadsheet\Worksheet\Sparkline\SparklineType;
+use PhpOffice\PhpSpreadsheet\Worksheet\PivotTable\PivotTable;
 
 class Worksheet
 {
@@ -138,6 +139,13 @@ class Worksheet
      * @var ArrayObject<int, SparklineGroup>
      */
     private ArrayObject $sparklineGroupCollection;
+
+    /**
+     * Collection of PivotTable objects.
+     *
+     * @var ArrayObject<int, PivotTable>
+     */
+    private ArrayObject $pivotTableCollection;
 
     /**
      * Worksheet title.
@@ -367,6 +375,9 @@ class Worksheet
         $this->tableCollection = new ArrayObject();
         // Sparkline group collection
         $this->sparklineGroupCollection = new ArrayObject();
+
+        // Pivot table collection
+        $this->pivotTableCollection = new ArrayObject();
     }
 
     /**
@@ -2236,6 +2247,80 @@ class Worksheet
     public function removeSparklineGroupCollection(): self
     {
         $this->sparklineGroupCollection = new ArrayObject();
+
+        return $this;
+    }
+
+    /**
+     * Get collection of PivotTables.
+     *
+     * @return ArrayObject<int, PivotTable>
+     */
+    public function getPivotTableCollection(): ArrayObject
+    {
+        return $this->pivotTableCollection;
+    }
+
+    /**
+     * Get collection of PivotTables (alias of getPivotTableCollection()).
+     *
+     * @return ArrayObject<int, PivotTable>
+     */
+    public function getPivotTables(): ArrayObject
+    {
+        return $this->pivotTableCollection;
+    }
+
+    /**
+     * Add a PivotTable to this worksheet.
+     *
+     * @return $this
+     */
+    public function addPivotTable(PivotTable $pivotTable): self
+    {
+        $pivotTable->setWorksheet($this);
+        $this->pivotTableCollection[] = $pivotTable;
+
+        return $this;
+    }
+
+    /**
+     * @return string[] array of PivotTable names
+     */
+    public function getPivotTableNames(): array
+    {
+        $pivotTableNames = [];
+
+        foreach ($this->pivotTableCollection as $pivotTable) {
+            $pivotTableNames[] = $pivotTable->getName();
+        }
+
+        return $pivotTableNames;
+    }
+
+    /**
+     * @param string $name the pivot table name to search
+     *
+     * @return null|PivotTable The pivot table from the collection, or null if not found
+     */
+    public function getPivotTableByName(string $name): ?PivotTable
+    {
+        $name = StringHelper::strToUpper($name);
+        foreach ($this->pivotTableCollection as $pivotTable) {
+            if (StringHelper::strToUpper($pivotTable->getName()) === $name) {
+                return $pivotTable;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Remove collection of PivotTables.
+     */
+    public function removePivotTableCollection(): self
+    {
+        $this->pivotTableCollection = new ArrayObject();
 
         return $this;
     }